// generated/notifications/Notification.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { CreatedMessageWithContent } from "./CreatedMessageWithContent";
import { SenderMetadata } from "./SenderMetadata";
import * as t from "io-ts";

/**
 * A received Notification.
 */

// required attributes
const NotificationR = t.interface({
  message: CreatedMessageWithContent,

  sender_metadata: SenderMetadata
});

// optional attributes
const NotificationO = t.partial({});

export const Notification = t.exact(
  t.intersection([NotificationR, NotificationO], "Notification")
);

export type Notification = t.TypeOf<typeof Notification>;